+++ src/TestSuite/TestCase.php
@@ -31,6 +31,6 @@ class TestCase extends CakeTestCase
     {
         parent::tearDown();
 
-        $this->deleteAllLogs();
+        safe_unlink_recursive(LOGS);
     }
 }